+++ Reach/test/utopia/reach/test/ReachInSwingTest.scala
@@ -2,10 +2,11 @@ package utopia.reach.test
 
 import utopia.flow.view.immutable.eventful.Fixed
 import utopia.flow.view.mutable.eventful.EventfulPointer
-import utopia.genesis.handling.KeyTypedListener
+import utopia.genesis.event.MouseButtonStateEvent
+import utopia.genesis.handling.{KeyTypedListener, MouseButtonStateListener}
 import utopia.genesis.view.GlobalKeyboardEventHandler
 import utopia.paradigm.color.Color
-import utopia.paradigm.color.ColorRole.Primary
+import utopia.paradigm.color.ColorRole.{Primary, Secondary}
 import utopia.paradigm.enumeration.Alignment.Center
 import utopia.paradigm.shape.shape2d.vector.point.Point
 import utopia.paradigm.shape.shape2d.vector.size.Size
@@ -20,9 +21,9 @@ import javax.swing.{JFrame, JPanel, WindowConstants}
   *
   * What you should see:
   *     - A window with a yellow window, opening at the center of the screen
-  *     - A red label somewhere in the upper left area of the window
-  *     - A black label at the center of the red label, reading "test"
+  *     - A black label at the center of an orange label, reading "test"
   *     - Whenever you type a character, that should modify the displayed text
+  *     - Whenver you click the label, that should change window color
   *
   * @author Mikko Hilpinen
   * @since 27.7.2023, v1.1
@@ -44,6 +45,7 @@ object ReachInSwingTest extends App
 			textPointer.value += char
 	}
 	
+	start()
 	window.setVisible(true)
 	
 	println("Done!")
@@ -53,9 +55,14 @@ object ReachInSwingTest extends App
 	
 	private class TestWindow extends JFrame()
 	{
-		val panel = new JPanel(null)
+		private val clicksCounter = EventfulPointer(0)
 		
-		panel.setBackground(Color.yellow.toAwt)
+		private val panel = new JPanel(null)
+		
+		clicksCounter.addContinuousListenerAndSimulateEvent(-1) { clicks =>
+			val color = if (clicks.newValue % 2 == 0) Color.yellow else Color.cyan
+			panel.setBackground(color.toAwt)
+		}
 		
 		setContentPane(panel)
 		setSize(300, 200)
@@ -63,15 +70,31 @@ object ReachInSwingTest extends App
 		setLocationRelativeTo(null)
 		setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE)
 		
-		val canvas = ReachCanvas.forSwing(Fixed(Color.transparentBlack), TestCursors.cursors, disableFocus = true) { hierarchy =>
-			AlignFrame(hierarchy).withContext(baseContext.against(Color.red).forTextComponents)(Center)
+		private val canvas = ReachCanvas.forSwing(baseContext.actorHandler, Fixed(Color.transparentBlack), TestCursors.cursors,
+			disableFocus = true) { hierarchy =>
+			AlignFrame(hierarchy).withContext(baseContext.against(Color.yellow).forTextComponents)(Center)
+				.withBackground(Secondary)
 				.build(ViewTextLabel) { labelF =>
-					labelF.withBackground(Primary)(textPointer)
+					val label = labelF.withBackground(Primary)(textPointer)
+					label.addMouseButtonListener(MouseButtonStateListener(MouseButtonStateEvent.leftPressedFilter) { event =>
+						if (label.bounds.contains(event.mousePosition))
+							clicksCounter.update { _ + 1 }
+						else
+							println("Click outside")
+						println(s"\tMouse: ${event.mousePosition}")
+						println(s"\tlabel: ${label.bounds}")
+						None
+					})
+					label
 				}
-		}.parent
-		canvas.position = Point(20, 20)
+		}
+		canvas.position = Point(20, 100)
 		canvas.size = Size(200, 50)
 		canvas.updateLayout()
+		canvas.child.addMouseButtonListener(MouseButtonStateListener(MouseButtonStateEvent.leftPressedFilter) { event =>
+			println(s"Align frame Mouse: ${event.mousePosition}")
+			None
+		})
 		
 		canvas.attachmentPointer.addListenerAndSimulateEvent(false) { e =>
 			if (e.newValue)