+++ Genesis/Changes.md
@@ -5,6 +5,8 @@
 - Added abstract function `.shade` to **ImageLike**
   - Only causes build errors within custom subclasses
 - Deleted functions and classes that were deprecated at v3.0 or earlier
+### Bugfixes
+- **MouseEventGenerator** that's directly attached to a **java.awt.Window** now accounts for the window insets
 ### New methods
 - **Image**
   - Added `.withCanvasSize(Size)` function, which adds padding or crops the image in order to get it to the desired size